Add getApiUrl module
This module defines the functions than can be used to compose the
URL to be used to fetch weather data using Visual Crossing API.

Link: https://www.visualcrossing.com/
Documentation: https://www.visualcrossing.com/resources/documentation/weather-api/timeline-weather-api/

The getApiUrl() can be used to retrieve the default current, hourly
and daily 15-day forecast weather data.

The getApiUrlBasic() can be used to retrieve only the minimal info
about the current conditions: datetime, temp, conditions, icon.

In both cases, the unit group is included in the link, and takes
the default 'metric' value.

Also, the location gets encoded. That is, the 'location' component
can be any string, with spaces, and so on.