Add renderSelectLocationPage module
This modules defines the renderSelectLocationPage(parentDiv)
function that creates a div showing an search input to enter a
location and renders it as a child of parentDiv.

A 'change' event listener has been added to such input element, so
that when the user press enter, the weather data for that locaton
are fetched. As of now, such data are only displayed in the console.

No error handling has been done yet.

Note that the getData() funcion has been modified to return the
Promise from fetch, in order to be able to access the data when
calling it (using getData(...).then(...)).