Setup basic renderWeatherDataPage module
The module renderWeatherDataPage defines the
renderWeatherDataPage(parentDiv, data) function that can be used
to display the data fetched from the weather API (and are passed
through the 'data' parameter). A div is created, which is set as
a child of the parentDiv.

We will initialy focus on properly displaying this page, rather
than the actual content, so only the location is inserted, right
now.