Setup PubSub to display the weather data when needed
PubSub has been used to trigger the rendering of valid weather
results. The subscription to the "RENDER WEATHER DATA" token has
been temporarly done in the index.js function, although that part
needs to be rearranged. When this token is published, the
renderWeatherDataPage function gets properly called.

Note that in the renderSelectLocationPage, a "RENDER WEATHER DATA"
token is published, along with the fetched data, when this needs
to be displayed.