Add different background when showing a .weather-data-page view
Right now, a sample linear-gradient background is used, which is defined
in the --weather-data-page-bg CSS variable.

The idea which is to be implemented is to change this background based
on the time of the day and the weather conditions, reflecting somehow
the color of the sky in that conditions.

Eg, in clear conditions, this background should appear light blue
during the day, and dark blue during the night. On sunset and sunrise
it should have red / violet shades.

When the cloud cover is high, this background should appear darker, while
if it is raining or the visibility is reduced (eg, there is fog), colors
should be washed out.