Set --weather-data-page-bg CSS variable via js
This is done each time a .weather-data-page is rendered. Right now,
the value is determined only based on day/night current condition, and
two solid colors are used.

In JS, colors will be represented as an array [R,G,B] in the RGB space.
getCssRgb() function is used to convert it in the CSS format.