Adjust RGB color to weather conditions in applyDynamicBackground
To adjust  a color based on weather conditions, we work on HSV color
space. The system is highly simplified, and the following observations
are considered:

- high cloud cover makes the sky appear darker and washed out
- precipitation makes the sky appear washed out
- low visibility makes the sky appear washed out

These are achieved as follows:

- to make a color darker, reduce its value in hsv space
- to make a color washed out, reduce its saturation in hsv space

For the saturation reduction, we just consider the weather factor that
has the biggest effect

In both cases, we don't want the color to appear too black or white,
respectively, so if we have a given reduction (factor) we both scale it
by a given reductionScalingFactor and cap it to a given maxReduction.
These values can be, in general, different between s and v computation.