Define sample day gradients
An array of sample gradients are defined which resemble the color of the
sky at different times of the day, from midnight to midnight of the next
day.

It assumes clear sky conditions (eg, no cloud cover, no precipitation
and maximum visibility).

Also, there are fixed times between stops: if there are 24 stops, there
is a color for each hour.

The sample gradient values are from: https://codepen.io/zessx/pen/kyEEBK

THese will be used to get the color of the background depending on the
time in which data refer to.