Install @date-fns/tz package
This will be used to take into account timezones for the dates used in
the app. Note that the timezone is provided by the Visual Crossing API.

The problem is that times are provided both as string and time epoch.
Dates objects have been created using new Date(epoch*1000), to have more
flexibility and be more useful. However, when using functions like
format from date-fns, ote method Date.getMinutes(), the timezone was
assumed to be the one from the local system, which leads to error in the
visualization of the dates.

With this package, it is possible to specify the timezone, eg, using:
format(dateObj,formatStr,{in: tz(timezoneStr)}), where timezoneStr is,
eg, "America/New_York".