Add getCurrentRgbGradient to set the background color based on the current time
The background gradient is obtained based on the gradient array from
sampleDayGradients module. This function assumes clear sky conditions,
and the timezone is taken into account, too.

Note that right now we assume that the gradient array has exactly 24
items (one for each hour).

Also, note that gradient array has 'fixed' sunrise and sunset moments.
In reality, each location might have different sunrise and sunset
hours, which results in a different background to be displayed.

To fix these issues, the gradient array needs to be manipulated, which
will be done next.