Preprocess the gradients array in sampleDayGradients before using it
This preprocessing consists in explicitly defining the 0% and 100% stops
if they are missing, and then, after having determined all the stop
percentages among all the gradients, determining the stop color for each
of these stop percentages, for each gradient. This is essentially
achieved through linear interpolation between existing stop values.
Indeed, recall that these stop values refer to a linear gradient.

In this way, even if we have a discrete number of stop gradients in the
gradient array, we can achieve the correct color for every hour of the
day (although this is not implemented yet).

Eg, if we have 24 stop values are we are in minute 567, we could get the
correct gradient interpolating between the sample gradients at indices 9
and 10 (as 567/60*24 = 9.45).

As an additional step, this gradients array should be stretched or
expanded based on sunrise and sunset hours.