Add interpolateGradient to interpolate between two gradients
We continue assuming the gradient from sampleDayGradient has 24 items,
each one associated with an hour.

Let's say we need the background for time 12.45. Now, we have to get the
sample gradients at indices 12 and 13, and interpolate with a factor of
0.75.

testDynamicBackground() now shows a continuous gradient during the day,
which is limited by the timeslot size.

Note that sunrise and sunset are not taken into account, yet.