Update getCurrentRgbGradient to consider current time, sunrise and sunset
The gradients from sampleDayGradients refer to some specific day,
where the sunrise and sunset time is set.

Now, in getCurrentRgbGradient, the current time is used to
determine whether it is late night, daytime or early night.
Depending on this, a different part of the gradients array is to
be considered, based on the sunset and sunrise time. That part of
the gradients array is stretched or expanded based on the relative
position of the current time inside either the day or the night.

Two gradients and a fraction are obtained, and the final gradient
is obtained through interpolation.