Add toggle bookmarked button in weather data page view
You can bookmark the current location or removed it from the
bookmarked locations.

Note that the bookmark buttons in the 'select location' page does
not update when a bookamrk is set or unset. While this could be
easily be done using pubsub, it will not be implemented, since that
'select' bookmark page will be a standalone page in the final
version, which will show the correct data.