Add showWeatherDataForCurrentPosition() function to use the current position as location
The current position is retrieved through the navigator.geolocation
function.

Note that this must be supported by the browser, and the user must have
given the 'Geolocation' permission.

See: https://developer.mozilla.org/en-US/docs/Web/API/Geolocation

Temporarily, the weather for the current position is shown in the
initial view, for testing.