Make the daily weather div buttons to show details for the day
Clicking on daily weather div li items now shows a detailed page with
information about that day. You can go back to the page showing the
current conditions with the back button, and select a different day on
that page, too.

The style is not completed yet, although most style is the same as the
weather data page.